Rename sql_formatter, add doc and examples
