+++ datafusion/physical-expr/src/expressions/cast.rs
@@ -240,7 +240,7 @@ pub fn cast(
 mod tests {
     use super::*;
 
-    use crate::{expressions::column::col, utils::sql_formatter};
+    use crate::expressions::column::col;
 
     use arrow::{
         array::{
@@ -251,6 +251,7 @@ mod tests {
         datatypes::*,
     };
     use datafusion_common::assert_contains;
+    use datafusion_physical_expr_common::physical_expr::fmt_sql;
 
     // runs an end-to-end test of physical type cast
     // 1. construct a record batch with a column "a" of type A
@@ -783,7 +784,7 @@ mod tests {
         let expr = cast(col("a", &schema)?, &schema, Int64)?;
         let display_string = expr.to_string();
         assert_eq!(display_string, "CAST(a@0 AS Int64)");
-        let sql_string = sql_formatter(expr.as_ref()).to_string();
+        let sql_string = fmt_sql(expr.as_ref()).to_string();
         assert_eq!(sql_string, "CAST(a AS Int64)");
 
         // Test string casting
@@ -791,7 +792,7 @@ mod tests {
         let expr = cast(col("b", &schema)?, &schema, Int32)?;
         let display_string = expr.to_string();
         assert_eq!(display_string, "CAST(b@0 AS Int32)");
-        let sql_string = sql_formatter(expr.as_ref()).to_string();
+        let sql_string = fmt_sql(expr.as_ref()).to_string();
         assert_eq!(sql_string, "CAST(b AS Int32)");
 
         Ok(())
