+++ datafusion/physical-expr/src/expressions/is_not_null.rs
@@ -120,13 +120,13 @@ pub fn is_not_null(arg: Arc<dyn PhysicalExpr>) -> Result<Arc<dyn PhysicalExpr>> 
 mod tests {
     use super::*;
     use crate::expressions::col;
-    use crate::utils::sql_formatter;
     use arrow::array::{
         Array, BooleanArray, Float64Array, Int32Array, StringArray, UnionArray,
     };
     use arrow::buffer::ScalarBuffer;
     use arrow::datatypes::*;
     use datafusion_common::cast::as_boolean_array;
+    use datafusion_physical_expr_common::physical_expr::fmt_sql;
 
     #[test]
     fn is_not_null_op() -> Result<()> {
@@ -213,7 +213,7 @@ mod tests {
         let expr = is_not_null(col("my_union", &schema).unwrap()).unwrap();
         let display_string = expr.to_string();
         assert_eq!(display_string, "my_union@0 IS NOT NULL");
-        let sql_string = sql_formatter(expr.as_ref()).to_string();
+        let sql_string = fmt_sql(expr.as_ref()).to_string();
         assert_eq!(sql_string, "my_union IS NOT NULL");
 
         Ok(())
