+++ datafusion/physical-expr/src/expressions/is_null.rs
@@ -119,13 +119,13 @@ pub fn is_null(arg: Arc<dyn PhysicalExpr>) -> Result<Arc<dyn PhysicalExpr>> {
 mod tests {
     use super::*;
     use crate::expressions::col;
-    use crate::utils::sql_formatter;
     use arrow::array::{
         Array, BooleanArray, Float64Array, Int32Array, StringArray, UnionArray,
     };
     use arrow::buffer::ScalarBuffer;
     use arrow::datatypes::*;
     use datafusion_common::cast::as_boolean_array;
+    use datafusion_physical_expr_common::physical_expr::fmt_sql;
 
     #[test]
     fn is_null_op() -> Result<()> {
@@ -224,7 +224,7 @@ mod tests {
         let expr = is_null(col("a", &schema)?).unwrap();
         let display_string = expr.to_string();
         assert_eq!(display_string, "a@0 IS NULL");
-        let sql_string = sql_formatter(expr.as_ref()).to_string();
+        let sql_string = fmt_sql(expr.as_ref()).to_string();
         assert_eq!(sql_string, "a IS NULL");
 
         Ok(())
