+++ datafusion/physical-expr/src/expressions/try_cast.rs
@@ -154,7 +154,6 @@ pub fn try_cast(
 mod tests {
     use super::*;
     use crate::expressions::col;
-    use crate::utils::sql_formatter;
     use arrow::array::{
         Decimal128Array, Decimal128Builder, StringArray, Time64NanosecondArray,
     };
@@ -165,6 +164,7 @@ mod tests {
         },
         datatypes::*,
     };
+    use datafusion_physical_expr_common::physical_expr::fmt_sql;
 
     // runs an end-to-end test of physical type cast
     // 1. construct a record batch with a column "a" of type A
@@ -589,7 +589,7 @@ mod tests {
         let expr = try_cast(col("a", &schema)?, &schema, DataType::Int64)?;
         let display_string = expr.to_string();
         assert_eq!(display_string, "TRY_CAST(a@0 AS Int64)");
-        let sql_string = sql_formatter(expr.as_ref()).to_string();
+        let sql_string = fmt_sql(expr.as_ref()).to_string();
         assert_eq!(sql_string, "TRY_CAST(a AS Int64)");
 
         // Test string casting
@@ -597,7 +597,7 @@ mod tests {
         let expr = try_cast(col("b", &schema)?, &schema, DataType::Int32)?;
         let display_string = expr.to_string();
         assert_eq!(display_string, "TRY_CAST(b@0 AS Int32)");
-        let sql_string = sql_formatter(expr.as_ref()).to_string();
+        let sql_string = fmt_sql(expr.as_ref()).to_string();
         assert_eq!(sql_string, "TRY_CAST(b AS Int32)");
 
         Ok(())
