--- datafusion/physical-expr/src/utils/sql_formatter.rs
@@ -1,35 +1,0 @@
-// Licensed to the Apache Software Foundation (ASF) under one
-// or more contributor license agreements.  See the NOTICE file
-// distributed with this work for additional information
-// regarding copyright ownership.  The ASF licenses this file
-// to you under the Apache License, Version 2.0 (the
-// "License"); you may not use this file except in compliance
-// with the License.  You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
-
-use std::fmt::{self, Formatter};
-
-use datafusion_physical_expr_common::physical_expr::PhysicalExpr;
-
-pub fn sql_formatter(expr: &dyn PhysicalExpr) -> impl fmt::Display + '_ {
-    struct Wrapper<'a> {
-        expr: &'a dyn PhysicalExpr,
-    }
-
-    impl fmt::Display for Wrapper<'_> {
-        fn fmt(&self, f: &mut Formatter) -> fmt::Result {
-            self.expr.fmt_sql(f)?;
-            Ok(())
-        }
-    }
-
-    Wrapper { expr }
-}
